+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+* The `HttpLoader::retryCachedErrorResponses()` method now returns an instance of the new `RetryManager` class, providing the methods `only()` and `except()` that can be used to restrict retries to certain HTTP response status codes.
 
 ## [1.10.0] - 2024-08-05
 ### Added